refactor: stable compilation state
