+++ .idea/nx-angular-config.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="NxAngularConfigService" workspaceLocation="file://$PROJECT_DIR$/nx.json" />
+  <component name="NxAngularConfigService" workspaceLocation="file://$PROJECT_DIR$/nx.json">
+    <projects>
+      <project name="automation" file="file://$PROJECT_DIR$/libs/automation/project.json" />
+    </projects>
+  </component>
 </project>
