+++ apps/api/src/app/workflows-v2/dto/replyCallbackDto.ts
@@ -1,0 +1,11 @@
+import { IsBoolean, IsDefined, IsString } from 'class-validator';
+
+export class ReplyCallbackDto {
+  @IsBoolean()
+  @IsDefined()
+  active: boolean;
+
+  @IsString()
+  @IsDefined()
+  url: string;
+}
